+++ main/Main.hs
@@ -14,18 +14,20 @@ module Main where
 
 import           Prelude                hiding (log, readFile)
 
-import           Control.Arrow          ((&&&), (>>>))
+import           Control.Arrow          ((>>>))
 import           Control.Monad          (join, when)
 import           Control.Monad.Trans    (MonadIO, liftIO)
 import           Data.Foldable          (foldMap, for_)
 import           Data.Function          ((&))
+import           Data.IORef
 import           Data.List              (isSuffixOf, sortOn)
 import           Data.Text              (Text)
 import qualified Data.Text              as T
 
 import           Data.Text.IO
+
 import           Path.Internal          (toFilePath)
-import           Path.IO                (getCurrentDir, listDir, resolveDir')
+import           Path.IO                (listDir, resolveDir')
 import           System.Console.Pretty
 import           System.Exit
 import           System.IO              (IOMode (..), stderr, stdout, withFile)
@@ -113,7 +115,8 @@ pipeline (NixLinter {..}) combined = let
   in
     S.fromList files
     & walk
-    & S.filter (isSuffixOf ".nix")
+    & S.filter ((recursive -->) <$> isSuffixOf ".nix")
+    & S.map (\p -> if p == "-" then "/dev/stdin" else p)
     & aheadly . parseFiles
     & aheadly . (S.map (combined >>> S.fromList) >>> join)
 
@@ -150,7 +153,14 @@ runChecks (opts@NixLinter{..}) = do
 
       results = pipeline opts combined
 
-  noIssues <- S.null results
-  withOutHandle $ \handle -> S.mapM_ (liftIO . printer handle) results
+  -- We can't use the naive implementation here, because that opens the files
+  -- multiple times
+  withOutHandle $ \handle -> do
+    hasIssues <- newIORef False
+    runStream $ flip S.mapM results $ \result -> do
+      printer handle result
+      -- "Smuggle" the result out of the Streamly datatype
+      writeIORef hasIssues True
 
-  if noIssues then exitSuccess else exitFailure
+    hadIssues <- readIORef hasIssues
+    if hadIssues then exitFailure else exitSuccess