+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## ??? -- 2019-??-??
 * Move more internal strings from String to Text
 * Add descriptions of offenses
+* Add `nix-linter -` to lint the contents of STDIN.
 
 ## 0.2.0.0 -- 2018-12-19